Add testing session
New Testing Session is for testing and development of this library.
Whilst you can use an XPLANID cookie from a XPLAN session for testing
purposes, this interface will not be made public, so it is subject
to change and removal at any time.

For production, you MUST use an API key.