Version Bump: 1.1.5
Commit 5682e93bac0444350d18d7633c9de4e72525d732 fixed a bug that
appeared in 48493e6adc71f0271dcbbc051000110d8188c756

Namely, with the introduction of URL encoding (which I forgot earlier),
the _transaction Query parameter was now encoded when before it wasn't.

Which is why the peculiar problem with XPlan API wasn't caught earlier.
The only way to catch these bugs currently is to test the code with a
real XPLAN server