Version 1.1.7: New macro for testing, found glaring bugs as a result
New internal macro: with-dummy-call
This macro allows one to make an API call (or bulk calls) and see what
the resulting URL and body content would be without actually making a call.

This is useful for finding bugs in internal logic.

Such as incorrectly using :post method for non-authenticating calls,
when it should have been the method variable that is determined by
CLOS functions...

Also, it is easier to spot where ':single-parms-as-body T' was omitted
when using custom methods in URL's... (because they are ALL POSTing anyway
I can just send actual parameters in body as JSON)

Which means I patched lots of entry-points where this bug existed...