Initial Work on 2.0.0-next. Backwards incompatible
This is the start of the work on the next major version.
It is a big breakaway form how it was done in the 1.X major version.

No longer will entry-points be defined, instead calls will be made
something like
(call-xplan *sess* $"/entity/client-v3/~!clientid!/asset/~!assetid!" :get)
if using the fstring-reader READ macro.

That is, you now pass the URL yourself, filling in need URL parameters.

This may end up being a Macro, with a separation between controls
for the call and parameters actually passed to the server.

Still in early re-design phase.